(FM-7922) running the configuration to apply catalog to transport
Largily the code here will not work on its own, as there are outstanding PRs for the `init_puppet_target` and the `trusted_facts` that are required.

There is also some changes to the configurer in Puppet that have been made locally, that will be discussed later/groomed.

This is mainly to raise discussions - we still need to sort out the environment isolation and split out the Puppet setup code from the `plugins` as it might be best to have a puppet settings class that isolates the puppet settings and then calls the pluginsync, and then the configurer - but again this is best to discuss as a team and this PR is not trying to refactor the existing code.