(maint) fixing up the docker setup for executing catalogs
Since the execution of catalogs requires Puppet configuration, the original base image did not have the required dependencies and initialisation was failing.

This change extends the puppet-agent alpine image and runs the ACE service on top in order to ensure the relevant dependencies are met.

Part of this change also required updating of the certificate requirements for working both within a container and `locally`, which means existing container users will need to regenerate new certs and restart the puppetserver,

```
cd spec/
docker-compose up -d --build --force-recreate
```

Instructions on regenerating the cert has been added to the docker.md file in developer-docs.