(FM-7954) plugin cache purge for stale environments
This will use a timer task to check every 24 hours for environments that have not been used in the last 7 days in the plugins folder.

Since the individual environment folders will be updated when a catalog is executed then only environments which have had no agentless catalogs executed in the last 7 days will get purged.

This should help keep excess plugins down, and active environments will be handled by the plugin handler from the Puppet code, i.e. it handles removing and updating of files per request if there are changes.