(PE-27794) add a puma status endpoint
With this commit, a puma status endpoint is added to our app.

Since this endpoint is served by our app, there will always be one
thread in use in the result: that thread is the status request itself.

This is not true with the default puma status endpoint,
as it runs within a separate server ... as per puma/lib/puma/runner.rb.