(PE-29007) Update to bolt > 2.9.0
This commit updates ace-server to reflect changes to the bolt `Result` object. This also fixes a bug in error creation whereby previously the `Target` we attempt to reference `name` for will always be the only element in an array, the `name` method was attempted to be called on the `Array`, this has been updated to call the method on the `Target`.