(PE-30044) Remove harmful terminology from ace
This commit removes the `whitelist` references in ace-server and
replaces them with `allowlist`. This is part of the effort to remove
harmful terminology, and will configure ace to use the new `allowlist`
ace-server.conf key rather than the old key.

This commit also updates a couple specs to account for new conf keys and
log levels. There's now a `project-dirs` conf key, and Bolt has changed
the default log level from 'notice' to 'warn'.