(PE-29359) Paper over spec failures
Currently the tests for our Certless Indirector implementation set up some quite intricate mocks in unit tests. The backing implementation changed slightly and unsurprisingly the mocks no longer work as intended. Instead of investing in mocking, we know we intend to make changes in puppet to expose the catalog v4 endpoint. At this point we can simply short circuit the mocks a bit to limp along this implementation which will be superceeded on short order.