(PE-31689) Don't fork in run_task endpoint
This was originally added as part of
https://github.com/puppetlabs/ace/pull/12 with no context behind the
addition. Turns out it is an unnecessary efficiency, so we revert back
to the previous behavior of just calling executor.run_task and having
the request-handling thread run the task. This is OK since the run_task
endpoint only takes in one target.

Signed-off-by: Enis Inan <enis.inan@puppet.com>