(PE-35500) Execute a puppet run on a device with a timeout
Share a `execute_with_timeout` function with run_task. This executes a puppet run (which is happening in a fork) in a thread that can be timed out. Note that the configurer needs to modify an object that is passed in to it so this must be done in the same thread.