(PE-35500) Timeout on waiting for fork for catalog application
The first approach for timeouts was to fork, then spin up a thread to do the work of catalog application etc, then if that thread timed out raise an error. This proved to be an issue with all the state and object references that caused us to have process isolation for this in the first place. With this commit we timeout waiting for the fork that sets up the puppet run on a device. This is probably safer anyway in case the process had hung outside of the puppet run.