(PE-35500) Add ability to timeout a remote task
This commit allows the run_task endpoint to accept a timeout (in seconds). If the timeout is hit the task will effectively be abandoned on the target. This is accomplished by running the task execution in a thread and waiting for that thred for "timeout" seconds. If the thread has not completed by that amount of time we send a kill signal to the thread and raise a timeout error.