(PE-35500) Support timeout for execute_catalog
This commit updates the execute_catalog endpoint to accept a timeout parameter. The timeout parameter represents the number of seconds to wait before giving up on the puppet run. This is implemented by waiting "timeout" seconds for the sub-process forked for the puppet run. If the timeout is hit the sub-process will be killed.